Component 6 Final Build
This is the final build of Component 6, which is essentially a combination of Components 1, 3 and 5, that allow for the creation of a Subtotal of Fixed Costs to be used later in the program. After this Component has been finished, I will combine it with Component 5 to create Component 7, which finds the total cost of the project.