Component 8 Final Build
This is the final build of Component 8, which calculates the profit and revenue required for a given project. This Component was made with pieces of code from two sources, both of which are given credit in the code itself. The next Component that I will begin work on is Component 9, which calculates the Profit Per Unit.