Component 9 First Build
This is the first build of Component 9. Currently, it mostly functions, however an equation for the "practical price per unit" has not yet been decided on. When I next work on this program, I will find a suitable "p_p_p_unit" calculation, which will ensure that the Component is fit for use, after which I will begin on my Assembled Outcome.