Component 9 Final Build
This is the final build of Component 9. This Component gives the user both a calculated and realistic price per unit, that they may use in order to determine their price per unit. This component has had decomposition screenshots taken of it, as of now, and I am now prepared to begin work on the Assembled Outcome. This concludes the Components section of the code that I have made from this Fund Raising Calculator.