Assembled Outcome First Build
This is the first build of the Fund Raising Calculator's Assembled Outcome (named 10_FRC_assembled_outcome on repl.it). This build has incorporated elements of Components 1, 2, 4 and 4b, as 3 seems to have no practical use at this time. The next stage of the Assembled Outcome will be the implementation of Components 5 and 6 (as well as 7 possibly), which may be difficult due to the difference in variable names for both fixed and variable expenses.