Component 7b First Build
This is the first build of Component 7b, which uses the variable and fixed costs from Components 5 and 6, and combines them to find the total. This component aims to make this process more efficeint, and while currently, this is not the case, this Component will allow the Fund Raising Calculator to be more efficient in the long term. I will continue my work on this Component tomorrow.