Functional Build of Component 7b
This is the first semi-functional build of Component 7b, which aims to create a more efficient Component 7. This build, while functional, is difficult for users to understand without making the code more difficult to understand. For this reason, I will be contacting Ms. Gottschalk in order to seek assistance with this code.