Component 7b Final Build
This is the final build of Component 7b, which calculates the total cost of the expenses that are given by the user, in a more efficient way than the original Component 7. This Component will also more easily be implemented into the Assembled Outcome, which I will continue during the next available Computer Science session.