Assembled Outcome with Updated String Checker Version and Statement Generator
This build of the Assembled Outcome is the one that is used in the Components 8 and 9 Decomposition Screenshot. Assembled Outcome "b" uses an updated string checker, which will not be used in the final Assembled Outcome due to minor errors. The original version of the Assembled Outcome is almost prepared to be Usability Tested.