Assembled Outcome with Project Summary Added
This is the build of the Assembled Outcome that ends with a summary of the costs associated tiht the project. This build ends with a series of print statements that detail specific costs associated with the project. At this point in the development process, almost all of the Assembled Outcome criteria have been completed, meaning that the outcome is close to being prepared for Usability Testing.