Usability Tested Assembled Outcome with Complete Instructions and clearer wording
This build of the Assembled Outcome implements many aspects to make the program easier to use, specifically in regards to the changes made to improve usabiliy after Usability Testing. In this build, a comprehensive set of instructions have been added, alongside woring changes to the request for the "units_required" variable that makes the question more easily understandable. Tomorrow, I will assess the state of the Assembled Outcome, which is now very close to completion.